move docker-agnostic container logic out of the docker package
