+++ pkg/util/containers/filter_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2018 Datadog, Inc.
 
-package docker
+package containers
 
 import (
 	"testing"
