+++ pkg/util/containers/metrics/network.go
@@ -3,9 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2018 Datadog, Inc.
 
-// +build docker
+// +build linux
 
-package docker
+package metrics
 
 import (
 	"fmt"
@@ -27,26 +27,24 @@ func (ns ContainerNetStats) SumInterfaces() *InterfaceNetStats {
 	return sum
 }
 
-func collectNetworkStats(containerID string, pid int, networks []dockerNetwork) (ContainerNetStats, error) {
+// CollectNetworkStats retrieves the network statistics for a given pid.
+// The networks map allows to optionnaly map interface name to user-friendly
+// network names. If not found in the map, the interface name is used.
+func CollectNetworkStats(pid int, networks map[string]string) (ContainerNetStats, error) {
 	netStats := ContainerNetStats{}
 
 	procNetFile := hostProc(strconv.Itoa(int(pid)), "net", "dev")
 	if !pathExists(procNetFile) {
-		log.Debugf("Unable to read %s for container %s", procNetFile, containerID)
+		log.Debugf("Unable to read %s for pid %s", procNetFile, pid)
 		return netStats, nil
 	}
 	lines, err := readLines(procNetFile)
 	if err != nil {
-		log.Debugf("Unable to read %s for container %s", procNetFile, containerID)
+		log.Debugf("Unable to read %s for pid %s", procNetFile, pid)
 		return netStats, nil
 	}
 	if len(lines) < 2 {
 		return nil, fmt.Errorf("invalid format for %s", procNetFile)
-	}
-
-	nwByIface := make(map[string]dockerNetwork)
-	for _, nw := range networks {
-		nwByIface[nw.iface] = nw
 	}
 
 	// Format:
@@ -65,12 +63,11 @@ func collectNetworkStats(containerID string, pid int, networks []dockerNetwork) 
 
 		var stat *InterfaceNetStats
 
-		if nw, ok := nwByIface[iface]; ok {
-			stat = &InterfaceNetStats{NetworkName: nw.dockerName}
+		if nw, ok := networks[iface]; ok {
+			stat = &InterfaceNetStats{NetworkName: nw}
 		} else if iface == "lo" {
 			continue // Ignore loopback
 		} else {
-			log.Debugf("Container %s: interface %s does not match a network, tagging with interface name", containerID, iface)
 			stat = &InterfaceNetStats{NetworkName: iface}
 		}
 
@@ -86,4 +83,48 @@ func collectNetworkStats(containerID string, pid int, networks []dockerNetwork) 
 		netStats = append(netStats, stat)
 	}
 	return netStats, nil
+}
+
+// DetectNetworkDestinations lists all the networks available
+// to a given PID and parses them in NetworkInterface objects
+func DetectNetworkDestinations(pid int) ([]NetworkDestination, error) {
+	procNetFile := hostProc(strconv.Itoa(int(pid)), "net", "route")
+	if !pathExists(procNetFile) {
+		return nil, fmt.Errorf("%s not found", procNetFile)
+	}
+	lines, err := readLines(procNetFile)
+	if err != nil {
+		return nil, err
+	}
+	if len(lines) < 1 {
+		return nil, fmt.Errorf("empty network file %s", procNetFile)
+	}
+
+	destinations := make([]NetworkDestination, 0)
+	for _, line := range lines[1:] {
+		fields := strings.Fields(line)
+		if len(fields) < 8 {
+			continue
+		}
+		if fields[1] == "00000000" {
+			continue
+		}
+		dest, err := strconv.ParseUint(fields[1], 16, 32)
+		if err != nil {
+			log.Debugf("Cannot parse destination %q: %v", fields[1], err)
+			continue
+		}
+		mask, err := strconv.ParseUint(fields[7], 16, 32)
+		if err != nil {
+			log.Debugf("Cannot parse mask %q: %v", fields[7], err)
+			continue
+		}
+		d := NetworkDestination{
+			Interface: fields[0],
+			Subnet:    dest,
+			Mask:      mask,
+		}
+		destinations = append(destinations, d)
+	}
+	return destinations, nil
 }
