+++ pkg/util/containers/types.go
@@ -3,7 +3,11 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-package docker
+package containers
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/util/containers/metrics"
+)
 
 // Expose container states
 const (
@@ -15,7 +19,20 @@ const (
 	ContainerDeadState       string = "dead"
 )
 
-// Container represents a single Docker container on a machine
+var (
+	// NullContainer is an empty container object that has
+	// default values for all fields including sub-fields.
+	// If new sub-structs are added to Container this must
+	// be updated.
+	NullContainer = &Container{
+		CPU:     &metrics.CgroupTimesStat{},
+		Memory:  &metrics.CgroupMemStat{},
+		IO:      &metrics.CgroupIOStat{},
+		Network: metrics.ContainerNetStats{},
+	}
+)
+
+// Container represents a single container on a machine
 // and includes Cgroup-level statistics about the container.
 type Container struct {
 	Type     string
@@ -34,12 +51,12 @@ type Container struct {
 	SoftMemLimit   uint64
 	MemLimit       uint64
 	CPUNrThrottled uint64
-	CPU            *CgroupTimesStat
-	Memory         *CgroupMemStat
-	IO             *CgroupIOStat
-	Network        ContainerNetStats
+	CPU            *metrics.CgroupTimesStat
+	Memory         *metrics.CgroupMemStat
+	IO             *metrics.CgroupIOStat
+	Network        metrics.ContainerNetStats
 	StartedAt      int64
 
 	// For internal use only
-	cgroup *ContainerCgroup
+	cgroup *metrics.ContainerCgroup
 }
