+++ pkg/util/docker/diagnosis.go
@@ -9,7 +9,6 @@ package docker
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/diagnose/diagnosis"
-
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
