+++ pkg/util/docker/host_metadata.go
