+++ pkg/util/docker/host_tags_test.go
