+++ pkg/util/docker/network_test.go
@@ -78,12 +78,10 @@ func TestFindDockerNetworks(t *testing.T) {
 
 	containerNetworks := make(map[string][]dockerNetwork)
 	for _, tc := range []struct {
-		pid         int
-		container   types.Container
-		routes, dev string
-		networks    []dockerNetwork
-		stat        ContainerNetStats
-		summedStat  *InterfaceNetStats
+		pid       int
+		container types.Container
+		routes    string
+		networks  []dockerNetwork
 	}{
 		// Host network mode
 		{
@@ -102,32 +100,11 @@ func TestFindDockerNetworks(t *testing.T) {
                 eth0    00000000    010011AC    0003    0   0   0   00000000    0   0   0
                 eth0    000011AC    00000000    0001    0   0   0   0000FFFF    0   0   0
             `),
-			dev: detab(`
-                Inter-|   Receive                                                |  Transmit
-                 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-                  eth0:    1345      10    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-                    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-            `),
 			networks: []dockerNetwork{{iface: "eth0", dockerName: "bridge"}},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "bridge",
-					BytesRcvd:   1345,
-					PacketsRcvd: 10,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   1345,
-				PacketsRcvd: 10,
-				BytesSent:   0,
-				PacketsSent: 0,
-			},
 		},
 		// No network mode, we treat this the same as host for now
 		{
-			pid: 1245,
+			pid: 1246,
 			container: types.Container{
 				ID: "test-find-docker-networks-nonetwork",
 				HostConfig: struct {
@@ -142,32 +119,11 @@ func TestFindDockerNetworks(t *testing.T) {
                 eth0    00000000    010011AC    0003    0   0   0   00000000    0   0   0
                 eth0    000011AC    00000000    0001    0   0   0   0000FFFF    0   0   0
             `),
-			dev: detab(`
-                Inter-|   Receive                                                |  Transmit
-                 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-                  eth0:    1345      10    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-                    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-            `),
 			networks: []dockerNetwork{{iface: "eth0", dockerName: "bridge"}},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "bridge",
-					BytesRcvd:   1345,
-					PacketsRcvd: 10,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   1345,
-				PacketsRcvd: 10,
-				BytesSent:   0,
-				PacketsSent: 0,
-			},
 		},
 		// Container network mode
 		{
-			pid: 1245,
+			pid: 1247,
 			container: types.Container{
 				ID: "test-find-docker-networks-container",
 				HostConfig: struct {
@@ -182,31 +138,10 @@ func TestFindDockerNetworks(t *testing.T) {
                 eth0    00000000    010011AC    0003    0   0   0   00000000    0   0   0
                 eth0    000011AC    00000000    0001    0   0   0   0000FFFF    0   0   0
             `),
-			dev: detab(`
-                Inter-|   Receive                                                |  Transmit
-                 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-                  eth0:    1345      10    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-                    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-            `),
 			networks: []dockerNetwork{{iface: "eth0", dockerName: "bridge"}},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "bridge",
-					BytesRcvd:   1345,
-					PacketsRcvd: 10,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   1345,
-				PacketsRcvd: 10,
-				BytesSent:   0,
-				PacketsSent: 0,
-			},
 		},
 		{
-			pid: 1245,
+			pid: 1248,
 			container: types.Container{
 				ID: "test-find-docker-networks-gateway",
 				HostConfig: struct {
@@ -227,32 +162,11 @@ func TestFindDockerNetworks(t *testing.T) {
                 eth0    00000000    010011AC    0003    0   0   0   00000000    0   0   0
                 eth0    000011AC    00000000    0001    0   0   0   0000FFFF    0   0   0
             `),
-			dev: detab(`
-                Inter-|   Receive                                                |  Transmit
-                 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-                  eth0:    1296      16    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-                    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-            `),
 			networks: []dockerNetwork{{iface: "eth0", dockerName: "eth0"}},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "eth0",
-					BytesRcvd:   1296,
-					PacketsRcvd: 16,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   1296,
-				PacketsRcvd: 16,
-				BytesSent:   0,
-				PacketsSent: 0,
-			},
 		},
 		// previous int32 overflow bug, now we parse uint32
 		{
-			pid: 1245,
+			pid: 1249,
 			container: types.Container{
 				ID: "test-find-docker-networks-gateway",
 				HostConfig: struct {
@@ -273,28 +187,7 @@ func TestFindDockerNetworks(t *testing.T) {
 		                eth0    00000000    FEFEA8C0    0003    0   0   0   00000000    0   0   0
 		                eth0    00FEA8C0    00000000    0001    0   0   0   00FFFFFF    0   0   0
 		            `),
-			dev: detab(`
-		                Inter-|   Receive                                                |  Transmit
-		                 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-		                  eth0:    1296      16    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-		                    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-		            `),
 			networks: []dockerNetwork{{iface: "eth0", dockerName: "eth0"}},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "eth0",
-					BytesRcvd:   1296,
-					PacketsRcvd: 16,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   1296,
-				PacketsRcvd: 16,
-				BytesSent:   0,
-				PacketsSent: 0,
-			},
 		},
 		// Multiple docker networks
 		{
@@ -318,141 +211,15 @@ func TestFindDockerNetworks(t *testing.T) {
 				eth0	000011AC	00000000	0001	0	0	0	0000FFFF	0	0	0
 				eth1	000012AC	00000000	0001	0	0	0	0000FFFF	0	0	0
             `),
-			dev: detab(`
-				Inter-|   Receive                                                |  Transmit
-				 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-				    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-				  eth0:     648       8    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-				  eth1:    1478      19    0    0    0     0          0         0      182       3    0    0    0     0       0          0`),
 			networks: []dockerNetwork{
 				{iface: "eth0", dockerName: "bridge"},
 				{iface: "eth1", dockerName: "test"},
-			},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "bridge",
-					BytesRcvd:   648,
-					PacketsRcvd: 8,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-				&InterfaceNetStats{
-					NetworkName: "test",
-					BytesRcvd:   1478,
-					PacketsRcvd: 19,
-					BytesSent:   182,
-					PacketsSent: 3,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   2126,
-				PacketsRcvd: 27,
-				BytesSent:   182,
-				PacketsSent: 3,
-			},
-		},
-		// Fallback to interface name if bridge is not in inspect (docker swarm bug)
-		{
-			pid: 5153,
-			container: types.Container{
-				ID: "test-find-docker-networks-multiple",
-				NetworkSettings: &types.SummaryNetworkSettings{
-					Networks: map[string]*dockernetwork.EndpointSettings{
-						"test": {
-							IPAddress: "172.18.0.1",
-						},
-					},
-				},
-			},
-			routes: detab(`
-				Iface	Destination	Gateway 	Flags	RefCnt	Use	Metric	Mask		MTU	Window	IRTT
-				eth0	00000000	010011AC	0003	0	0	0	00000000	0	0	0
-				eth0	000011AC	00000000	0001	0	0	0	0000FFFF	0	0	0
-				eth1	000012AC	00000000	0001	0	0	0	0000FFFF	0	0	0
-            `),
-			dev: detab(`
-				Inter-|   Receive                                                |  Transmit
-				 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-				    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-				  eth0:     648       8    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-				  eth1:    1478      19    0    0    0     0          0         0      182       3    0    0    0     0       0          0`),
-			networks: []dockerNetwork{
-				{iface: "eth1", dockerName: "test"},
-			},
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "eth0",
-					BytesRcvd:   648,
-					PacketsRcvd: 8,
-					BytesSent:   0,
-					PacketsSent: 0,
-				},
-				&InterfaceNetStats{
-					NetworkName: "test",
-					BytesRcvd:   1478,
-					PacketsRcvd: 19,
-					BytesSent:   182,
-					PacketsSent: 3,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   2126,
-				PacketsRcvd: 27,
-				BytesSent:   182,
-				PacketsSent: 3,
-			},
-		},
-		// Dumb error case to make sure we don't panic, fallback to interface name
-		{
-			pid: 5157,
-			container: types.Container{
-				ID: "test-find-docker-networks-errcase",
-				HostConfig: struct {
-					NetworkMode string `json:",omitempty"`
-				}{
-					NetworkMode: "isolated_nw",
-				},
-				NetworkSettings: &types.SummaryNetworkSettings{
-					Networks: map[string]*dockernetwork.EndpointSettings{
-						"isolated_nw": {
-							IPAddress: "172.18.0.1",
-						},
-						"eth0": {
-							IPAddress: "172.0.0.4/24",
-						},
-					},
-				},
-			},
-			routes:   detab(``),
-			networks: nil,
-			dev: detab(`
-                Inter-|   Receive                                                |  Transmit
-                 face |bytes    packets errs drop fifo frame compressed multicast|bytes    packets errs drop fifo colls carrier compressed
-                  eth0:    1111       2    0    0    0     0          0         0     1024      80    0    0    0     0       0          0
-                    lo:       0       0    0    0    0     0          0         0        0       0    0    0    0     0       0          0
-            `),
-			stat: ContainerNetStats{
-				&InterfaceNetStats{
-					NetworkName: "eth0",
-					BytesRcvd:   1111,
-					PacketsRcvd: 2,
-					BytesSent:   1024,
-					PacketsSent: 80,
-				},
-			},
-			summedStat: &InterfaceNetStats{
-				BytesRcvd:   1111,
-				PacketsRcvd: 2,
-				BytesSent:   1024,
-				PacketsSent: 80,
 			},
 		},
 	} {
 		t.Run("", func(t *testing.T) {
 			// Create temporary files on disk with the routes and stats.
 			err = dummyProcDir.add(filepath.Join(strconv.Itoa(int(tc.pid)), "net", "route"), tc.routes)
-			assert.NoError(t, err)
-			err = dummyProcDir.add(filepath.Join(strconv.Itoa(int(tc.pid)), "net", "dev"), tc.dev)
 			assert.NoError(t, err)
 
 			// Use the routes file and settings to get our networks.
@@ -463,12 +230,6 @@ func TestFindDockerNetworks(t *testing.T) {
 			resolveDockerNetworks(containerNetworks)
 			networks = containerNetworks[tc.container.ID]
 			assert.Equal(t, tc.networks, networks)
-
-			// And collect the stats on these networks.
-			stat, err := collectNetworkStats(tc.container.ID, tc.pid, networks)
-			assert.NoError(t, err)
-			assert.Equal(t, tc.stat, stat)
-			assert.Equal(t, tc.summedStat, stat.SumInterfaces())
 		})
 	}
 }
